# docs/swagger.yaml
basePath: /api/v2
definitions:
...
  api.MeasurementResponse:
    properties:
      created_at:
        type: string
      date:
        type: string
      ftp:
        type: number
      height:
        type: number
      id:
        type: integer
      max_heart_rate:
        type: number
      resting_heart_rate:
        type: number
      steps:
        type: integer
      updated_at:
        type: string
      user_id:
        type: integer
      weight:
        type: number
    type: object
...
  api.PaginatedResponse-api_MeasurementResponse:
    properties:
      errors:
        items:
          type: string
        type: array
      page:
        type: integer
      per_page:
        type: integer
      results:
        items:
          $ref: '#/definitions/api.MeasurementResponse'
        type: array
      total_count:
        type: integer
      total_pages:
        type: integer
    type: object
...
  api.ProfileSettings:
    properties:
      api_active:
        type: boolean
      api_key:
        type: string
      auto_import_directory:
        type: string
      language:
        type: string
      prefer_full_date:
        type: boolean
      preferred_units:
        $ref: '#/definitions/database.UserPreferredUnits'
      socials_disabled:
        type: boolean
      theme:
        type: string
      timezone:
        type: string
      totals_show:
        type: string
    type: object
...
  api.Response-any:
    properties:
      errors:
        items:
          type: string
        type: array
      results: {}
    type: object
...
  api.Response-api_MeasurementResponse:
    properties:
      errors:
        items:
          type: string
        type: array
      results:
        $ref: '#/definitions/api.MeasurementResponse'
    type: object
...
  api.Response-api_UserProfileResponse:
    properties:
      errors:
        items:
          type: string
        type: array
      results:
        $ref: '#/definitions/api.UserProfileResponse'
    type: object
...
  api.Response-map_string_string:
    properties:
      errors:
        items:
          type: string
        type: array
      results:
        $ref: '#/definitions/map_string_string'
    type: object
...
  api.UserProfileResponse:
    properties:
      active:
        type: boolean
      admin:
        type: boolean
      birthdate:
        type: string
      created_at:
        type: string
      id:
        type: integer
      language:
        type: string
      last_version:
        type: string
      name:
        type: string
      prefer_full_date:
        type: boolean
      preferred_units:
        $ref: '#/definitions/database.UserPreferredUnits'
      profile:
        $ref: '#/definitions/api.ProfileSettings'
      socials_disabled:
        type: boolean
      theme:
        type: string
      timezone:
        type: string
      updated_at:
        type: string
      username:
        type: string
    type: object
...
  database.UserPreferredUnits:
    properties:
      distance:
        description: The user's preferred distance unit
        type: string
      elevation:
        description: The user's preferred elevation unit
        type: string
      height:
        description: The user's preferred height unit
        type: string
      speed:
        description: The user's preferred speed unit
        type: string
      weight:
        description: The user's preferred weight unit
        type: string
    type: object
...
  map_string_string:
    additionalProperties:
      type: string
    type: object
info:
  contact: {}
  description: Workout Tracker HTTP API (v2).
  title: Workout Tracker API
  version: "2.0"
paths:
...
  /measurements:
    get:
      parameters:
      - description: Page
        in: query
        name: page
        type: integer
      - description: Per page
        in: query
        name: per_page
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/api.PaginatedResponse-api_MeasurementResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/api.Response-any'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/api.Response-any'
      security:
      - ApiKeyAuth: []
      - ApiKeyQuery: []
      - CookieAuth: []
      summary: List measurements
      tags:
      - measurements
    post:
      consumes:
      - application/json
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/api.Response-api_MeasurementResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/api.Response-any'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/api.Response-any'
      security:
      - ApiKeyAuth: []
      - ApiKeyQuery: []
      - CookieAuth: []
      summary: Create or update measurement
      tags:
      - measurements
  /measurements/{date}:
    delete:
      parameters:
      - description: Date (YYYY-MM-DD)
        in: path
        name: date
        required: true
        type: string
      produces:
      - application/json
      responses:
        "204":
          description: No Content
          schema:
            type: string
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/api.Response-any'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/api.Response-any'
      security:
      - ApiKeyAuth: []
      - ApiKeyQuery: []
      - CookieAuth: []
      summary: Delete measurement by date
      tags:
      - measurements
  /profile:
    get:
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/api.Response-api_UserProfileResponse'
      security:
      - ApiKeyAuth: []
      - ApiKeyQuery: []
      - CookieAuth: []
      summary: Get profile
      tags:
      - profile
    put:
      consumes:
      - application/json
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/api.Response-api_UserProfileResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/api.Response-any'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/api.Response-any'
      security:
      - ApiKeyAuth: []
      - ApiKeyQuery: []
      - CookieAuth: []
      summary: Update profile
      tags:
      - profile
  /profile/refresh-workouts:
    post:
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/api.Response-map_string_string'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/api.Response-any'
      security:
      - ApiKeyAuth: []
      - ApiKeyQuery: []
      - CookieAuth: []
      summary: Refresh all workouts
      tags:
      - profile
  /profile/reset-api-key:
    post:
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/api.Response-map_string_string'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/api.Response-any'
      security:
      - ApiKeyAuth: []
      - ApiKeyQuery: []
      - CookieAuth: []
      summary: Reset API key
      tags:
      - profile
...
securityDefinitions:
  ApiKeyAuth:
    in: header
    name: Authorization
    type: apiKey
  ApiKeyQuery:
    in: query
    name: api-key
    type: apiKey
  CookieAuth:
    in: header
    name: Cookie
    type: apiKey
swagger: "2.0"